EventsView: avoid narrowing conversion
getRunningTime() returns a long

+++ runtime/bundles/org.eclipse.core.tools/src/org/eclipse/core/tools/runtime/EventsView.java
@@ -14,17 +14,36 @@
 
 package org.eclipse.core.tools.runtime;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
 import org.eclipse.core.runtime.PerformanceStats;
-import org.eclipse.core.tools.*;
-import org.eclipse.jface.action.*;
+import org.eclipse.core.tools.CopyStructuredSelectionAction;
+import org.eclipse.core.tools.Messages;
+import org.eclipse.core.tools.TableSelectionProviderDecorator;
+import org.eclipse.core.tools.TableWithTotalView;
+import org.eclipse.jface.action.Action;
+import org.eclipse.jface.action.IToolBarManager;
+import org.eclipse.jface.action.MenuManager;
 import org.eclipse.jface.resource.ImageDescriptor;
-import org.eclipse.jface.viewers.*;
+import org.eclipse.jface.viewers.ColumnLayoutData;
+import org.eclipse.jface.viewers.ColumnPixelData;
+import org.eclipse.jface.viewers.ColumnWeightData;
+import org.eclipse.jface.viewers.IColorProvider;
+import org.eclipse.jface.viewers.ITableLabelProvider;
+import org.eclipse.jface.viewers.ITreeContentProvider;
+import org.eclipse.jface.viewers.LabelProvider;
+import org.eclipse.jface.viewers.TableViewer;
+import org.eclipse.jface.viewers.Viewer;
+import org.eclipse.jface.viewers.ViewerComparator;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.Image;
-import org.eclipse.swt.widgets.*;
+import org.eclipse.swt.widgets.Composite;
+import org.eclipse.swt.widgets.Control;
+import org.eclipse.swt.widgets.Display;
+import org.eclipse.swt.widgets.Menu;
 import org.eclipse.ui.IActionBars;
 
 /**
@@ -185,7 +204,7 @@ public class EventsView extends TableWithTotalView {
 		String[] totals = new String[getColumnHeaders().length];
 		int count = 0;
 		int events = 0;
-		int time = 0;
+		long time = 0;
 		if (!iter.hasNext()) {
 			Object[] elements = ((ITreeContentProvider) viewer.getContentProvider()).getElements(viewer.getInput());
 			@SuppressWarnings({ "rawtypes", "unchecked" })
